+++ diff/diff.go
@@ -7,7 +7,7 @@ import (
 
 	lflags "github.com/launchdarkly/find-code-references-in-pull-request/flags"
 	i "github.com/launchdarkly/find-code-references-in-pull-request/ignore"
-	diff_util "github.com/launchdarkly/find-code-references-in-pull-request/internal/util/diff_util"
+	diff_util "github.com/launchdarkly/find-code-references-in-pull-request/internal/utils/diff_util"
 	"github.com/launchdarkly/ld-find-code-refs/v2/aliases"
 	lsearch "github.com/launchdarkly/ld-find-code-refs/v2/search"
 	"github.com/sourcegraph/go-diff/diff"
