+++ internal/utils/diff_util/operation.go
